requirements.txt
aws_cdk.aws_lambda
aws_cdk.aws_events
aws_cdk.aws_sqs
aws_cdk.aws_events_targets
